refactor: extract helpers/pricing.py (Phase 6.1 — first helper move)
Start of the helpers/ migration flagged in CLAUDE.md: "Helpers will migrate
to helpers/ over time." Moves the two pure provider-classification helpers
out of dashboard.py's module-level bag:

  _provider_from_model(model_name)       — strict prefix/substring match
                                            (used for pricing table lookups)
  _infer_provider_from_model(model_name) — looser substring match for
                                            display-only provider labels

Both are pure functions with zero module-level dependencies, so this is
a mechanical move. Re-exported from dashboard.py at module level so
existing `_d._provider_from_model(...)` calls in routes/*.py keep working
without code changes. Over time, route modules will import from helpers/
directly and the re-export shim will be retired.

## Call sites validated

  routes/overview.py:316    _d._infer_provider_from_model(...)
  routes/components.py:933  _d._provider_from_model(...)
  routes/usage.py:799       _d._provider_from_model(...)

## E2E

- helpers/pricing.py unit-tested: 6 assertions pass (openai/anthropic/
  xai/google/unknown prefixes, llama → local/other, empty → unknown)
- 35/35 API endpoints return 200 (every tab across all 13 route modules +
  main HTML page)
- Zero tracebacks in server log

## Sizes

  dashboard.py       : -34 lines (two function bodies removed; 7-line
                                  re-export import added; net -27)
  helpers/__init__.py: new (0 lines)
  helpers/pricing.py : new (55 lines with docstrings)
